liquer-pcv/requirements.txt
flask>=1.0.2
liquer-framework>=0.4.11
